Add files via upload
Please copy this file in the "presets" folder of Videomass.

These presets are based on NVENC encoding (NVIDIA accelerated video encoding).
All the presets take all streams (audio and video) and subtitles and put them in a new MKV file format.
The presets are based on video bitrate (2600kb) or quality setting (cq 23 - medium). The audio options are aac, mp3 (stereo only), vorbis and PCM. Also there are seperate presets for the audio to be normalized according to loudness. Finally there are presets where only the video or the audio is encoded and the other stream is copied (lossless).

The video bitrate of 2600kb is for me the best setting to achieve good quality and small size. 
The quality setting of 23 is better quality and again a sweet spot on quality/file size but will most of time lead to a bigger filesize.

Feel free to adjust the video and audio settings.